# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Containr API
  description: |
    Self-hosted PaaS on Proxmox - Railway-like platform for container orchestration and deployment.
    
    ## Overview
    Containr provides a comprehensive platform for deploying, managing, and scaling containerized applications on Proxmox infrastructure. This API allows you to manage projects, services, deployments, databases, and more through a RESTful interface.
    
    ## Authentication
    The API uses JWT-based authentication. Include your JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Base URL
    ```
    https://your-containr-domain.com/api/v1
    ```
  version: 1.0.0
  contact:
    name: Containr Documentation
    url: https://github.com/your-org/containr
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://your-containr-domain.com/api/v1
    description: Production server
paths:
...
  /ha/status:
    get:
      tags:
        - HA
      summary: HA status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Overall HA status (enabled, policies, active alerts)
          content:
            application/json:
              schema:
                type: object
  /ha/enable:
    post:
      tags:
        - HA
      summary: Enable HA management
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HA enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /ha/disable:
    post:
      tags:
        - HA
      summary: Disable HA management
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HA disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /ha/failover:
    post:
      tags:
        - HA
      summary: Trigger failover
      description: Manually trigger a failover evaluation across services with failover policies.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Failover triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /ha/failover/policies:
    get:
      tags:
        - HA
      summary: List failover policies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All failover policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: '#/components/schemas/FailoverPolicy'
    post:
      tags:
        - HA
      summary: Create or replace failover policy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailoverPolicy'
      responses:
        '201':
          description: Policy stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  policy:
                    $ref: '#/components/schemas/FailoverPolicy'
  /ha/failover/policies/{serviceId}:
    get:
      tags:
        - HA
      summary: Get failover policy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '200':
          description: Failover policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  policy:
                    $ref: '#/components/schemas/FailoverPolicy'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - HA
      summary: Update failover policy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailoverPolicy'
      responses:
        '200':
          description: Policy updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  policy:
                    $ref: '#/components/schemas/FailoverPolicy'
    delete:
      tags:
        - HA
      summary: Delete failover policy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '200':
          description: Policy deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /ha/alerts/rules:
    get:
      tags:
        - HA
      summary: List alert rules
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All alert rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertRule'
    post:
      tags:
        - HA
      summary: Add alert rule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRule'
      responses:
        '201':
          description: Alert rule created
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule:
                    $ref: '#/components/schemas/AlertRule'
  /ha/alerts/rules/{ruleId}:
    get:
      tags:
        - HA
      summary: Get alert rule
      security:
        - bearerAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert rule
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule:
                    $ref: '#/components/schemas/AlertRule'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - HA
      summary: Update alert rule
      security:
        - bearerAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRule'
      responses:
        '200':
          description: Alert rule updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule:
                    $ref: '#/components/schemas/AlertRule'
    delete:
      tags:
        - HA
      summary: Delete alert rule
      security:
        - bearerAuth: []
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert rule deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /ha/alerts/active:
    get:
      tags:
        - HA
      summary: Active alerts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Currently firing alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      type: object
  /ha/alerts/{alertId}/resolve:
    post:
      tags:
        - HA
      summary: Resolve alert
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /ha/notifiers:
    get:
      tags:
        - HA
      summary: List notifiers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configured alert notifiers
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifiers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notifier'
    post:
      tags:
        - HA
      summary: Add notifier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotifierRequest'
      responses:
        '201':
          description: Notifier created
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifier:
                    $ref: '#/components/schemas/Notifier'
  /ha/notifiers/{notifierId}:
    get:
      tags:
        - HA
      summary: Get notifier
      security:
        - bearerAuth: []
      parameters:
        - name: notifierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notifier
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifier:
                    $ref: '#/components/schemas/Notifier'
        '404':
          description: Notifier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - HA
      summary: Delete notifier
      security:
        - bearerAuth: []
      parameters:
        - name: notifierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notifier deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # Node agents — management (JWT-authenticated)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    FailoverPolicy:
      type: object
      required:
        - service_id
      properties:
        service_id:
          type: string
        enabled:
          type: boolean
        min_healthy_nodes:
          type: integer
        max_failures:
          type: integer
        failover_timeout:
          type: integer
          description: Nanoseconds
        recovery_timeout:
          type: integer
          description: Nanoseconds
        failover_strategy:
          type: string
        backup_nodes:
          type: array
          items:
            type: string
        health_check_config:
          type: object
...
    AlertRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        condition:
          type: object
          description: metric, operator, threshold, duration
        severity:
          type: string
          enum: [info, warning, critical]
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        notifiers:
          type: array
          items:
            type: string
        cooldown:
          type: integer
          description: Nanoseconds
    Notifier:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          description: webhook | slack | email | discord
        config:
          type: object
    NotifierRequest:
      type: object
      required:
        - type
        - config
      properties:
        id:
          type: string
        type:
          type: string
        config:
          type: object
...
  parameters:
...
    ServiceIdParam:
      name: serviceId
      in: path
      required: true
      description: Service ID
      schema:
        type: string
...